Refactor get_workflows()
