+++ refinery/galaxy_connector/galaxy_workflow.py
@@ -16,29 +16,6 @@ import core
 logger = logging.getLogger(__name__)
 
 
-class GalaxyWorkflow(object):
-    def __init__(self, name, identifier):
-        self.name = name
-        self.identifier = identifier
-        self.inputs = []
-
-    def __unicode__(self):
-        return self.name + " (" + self.identifier + "): " + str(
-            len(self.inputs)) + " inputs"
-
-    def add_input(self, workflow_input):
-        self.inputs.append(workflow_input)
-
-
-class GalaxyWorkflowInput(object):
-    def __init__(self, name, identifier):
-        self.name = name
-        self.identifier = identifier
-
-    def __unicode__(self):
-        return self.name + " (" + self.identifier + ")"
-
-
 # Helper functions
 def createBaseWorkflow(workflow_name):
     """Creates base template workflow"""
