+++ refinery/galaxy_connector/models.py
@@ -2,8 +2,6 @@ from django.db import models
 
 from bioblend import galaxy
 
-import galaxy_workflow
-
 
 class Instance(models.Model):
     base_url = models.CharField(max_length=2000)
@@ -19,22 +17,6 @@ class Instance(models.Model):
 
     def galaxy_connection(self):
         return galaxy.GalaxyInstance(url=self.base_url, key=self.api_key)
-
-    def get_complete_workflows(self):
-        connection = self.galaxy_connection()
-        workflows = []
-        for workflow_entry in connection.workflows.get_workflows():
-            workflow = galaxy_workflow.GalaxyWorkflow(workflow_entry['name'],
-                                                      workflow_entry['id'])
-            # get workflow inputs
-            workflow_inputs = connection.workflows.show_workflow(
-                workflow.identifier)['inputs']
-            for input_identifier, input_description in workflow_inputs.items():
-                workflow_input = galaxy_workflow.GalaxyWorkflowInput(
-                    input_description['label'], input_identifier)
-                workflow.add_input(workflow_input)
-            workflows.append(workflow)
-        return workflows
 
     def get_history_file_list(self, history_id):
         """Returns a list of dictionaries that contain the name, type, state
